binding
- refacto tes3.createObject
