+++ MWSE/TES3UtilLua.cpp
@@ -2236,92 +2236,7 @@ namespace mwse {
 			{
 				TES3::ObjectType::ObjectType objectType = getOptionalParam( params, "objectType", TES3::ObjectType::Invalid );
 
-				switch( objectType )
-				{
-				case TES3::ObjectType::Activator:
-					return createActivator( params );
-				case TES3::ObjectType::Alchemy:
-					throw std::invalid_argument{ "tes3.createObject: alchemy object type is not handled." };
-				case TES3::ObjectType::Ammo:
-					throw std::invalid_argument{ "tes3.createObject: ammunition object type is not handled." };
-				case TES3::ObjectType::Apparatus:
-					throw std::invalid_argument{ "tes3.createObject: apparatus object type is not handled." };
-				case TES3::ObjectType::Armor:
-					throw std::invalid_argument{ "tes3.createObject: armor object type is not handled." };
-				case TES3::ObjectType::Book:
-					throw std::invalid_argument{ "tes3.createObject: book object type is not handled." };
-				case TES3::ObjectType::Clothing:
-					throw std::invalid_argument{ "tes3.createObject: clothing object type is not handled." };
-				case TES3::ObjectType::Container:
-					throw std::invalid_argument{ "tes3.createObject: container object type is not handled." };
-				case TES3::ObjectType::Creature:
-					throw std::invalid_argument{ "tes3.createObject: creature object type is not handled." };
-				case TES3::ObjectType::Door:
-					throw std::invalid_argument{ "tes3.createObject: door object type is not handled." };
-				case TES3::ObjectType::Enchantment:
-					throw std::invalid_argument{ "tes3.createObject: enchantment object type is not handled." };
-				case TES3::ObjectType::Faction:
-					throw std::invalid_argument{ "tes3.createObject: faction object type is not handled." };
-				case TES3::ObjectType::Ingredient:
-					throw std::invalid_argument{ "tes3.createObject: ingredient object type is not handled." };
-				case TES3::ObjectType::LeveledCreature:
-					throw std::invalid_argument{ "tes3.createObject: leveledCreature object type is not handled." };
-				case TES3::ObjectType::LeveledItem:
-					throw std::invalid_argument{ "tes3.createObject: leveldItem object type is not handled." };
-				case TES3::ObjectType::Light:
-					throw std::invalid_argument{ "tes3.createObject: light object type is not handled." };
-				case TES3::ObjectType::Lockpick:
-					throw std::invalid_argument{ "tes3.createObject: lockPick object type is not handled." };
-				case TES3::ObjectType::Misc:
-					return createMiscItem( params );
-				case TES3::ObjectType::MobileCreature:
-					throw std::invalid_argument{ "tes3.createObject: mobileCreature object type is not handled." };
-				case TES3::ObjectType::MobileNPC:
-					throw std::invalid_argument{ "tes3.createObject: mobileNpc object type is not handled." };
-				case TES3::ObjectType::NPC:
-					throw std::invalid_argument{ "tes3.createObject: npc object type is not handled." };
-				case TES3::ObjectType::Probe:
-					throw std::invalid_argument{ "tes3.createObject: probe object type is not handled." };
-				case TES3::ObjectType::Reference:
-					throw std::invalid_argument{ "tes3.createObject: reference object type is not handled." };
-				case TES3::ObjectType::Repair:
-					throw std::invalid_argument{ "tes3.createObject: repair object type is not handled." };
-				case TES3::ObjectType::Script:
-					throw std::invalid_argument{ "tes3.createObject: script object type is not handled." };
-				case TES3::ObjectType::SoundGenerator:
-					throw std::invalid_argument{ "tes3.createObject: soundGenerator object type is not handled." };
-				case TES3::ObjectType::Spell:
-					throw std::invalid_argument{ "tes3.createObject: spell object type is not handled." };
-				case TES3::ObjectType::Static:
-					return createStatic( params );
-				case TES3::ObjectType::Weapon:
-					throw std::invalid_argument{ "tes3.createObject: weapon object type is not handled." };
-				case TES3::ObjectType::Invalid:
-				case TES3::ObjectType::Birthsign:
-				case TES3::ObjectType::Bodypart:
-				case TES3::ObjectType::Cell:
-				case TES3::ObjectType::Class:
-				case TES3::ObjectType::Dialogue:
-				case TES3::ObjectType::DialogueInfo:
-				case TES3::ObjectType::GameSetting:
-				case TES3::ObjectType::Global:
-				case TES3::ObjectType::Land:
-				case TES3::ObjectType::LandTexture:
-				case TES3::ObjectType::MagicEffect:
-				case TES3::ObjectType::MagicSourceInstance:
-				case TES3::ObjectType::MobileObject:
-				case TES3::ObjectType::MobilePlayer:
-				case TES3::ObjectType::MobileProjectile:
-				case TES3::ObjectType::Quest:
-				case TES3::ObjectType::PathGrid:
-				case TES3::ObjectType::Race:
-				case TES3::ObjectType::Region:
-				case TES3::ObjectType::Skill:
-				case TES3::ObjectType::Sound:
-				case TES3::ObjectType::TES3:
-				default:
-					throw std::invalid_argument{ "tes3.createObject: Invalid objectType parameter specified." };
-				}
+				return ObjectCreatorFactory{}.getObjectCreator( objectType )->create( params );
 			};
 
 			state["tes3"]["setDestination"] = [](sol::table params) {
